--- a/compiler/src/Language/Mimsa/Interpreter/App.hs
+++ b/compiler/src/Language/Mimsa/Interpreter/App.hs
@@ -1,38 +1,48 @@
 module Language.Mimsa.Interpreter.App (interpretApp) where
 
-import Language.Mimsa.Core
-import Language.Mimsa.Interpreter.Monad
+import Language.Mimsa.Logging
+import  Language.Mimsa.Interpreter.Monad
+import qualified Language.Mimsa.Interpreter.HOASExpr as HOAS
+import Language.Mimsa.Interpreter.ToHOAS
 import Language.Mimsa.Interpreter.Types
-import Language.Mimsa.Types.Interpreter.Stack
-
-varFromIdent :: Identifier var ann -> var
-varFromIdent (Identifier _ var) = var
+import Language.Mimsa.Core
 
 interpretApp ::
-  (Ord var, Eq ann) =>
-  InterpretFn var ann ->
-  ExprData var ann ->
-  InterpretExpr var ann ->
-  InterpretExpr var ann ->
-  InterpreterM var ann (InterpretExpr var ann)
-interpretApp interpretFn ann myFn value =
-  case myFn of
-    (MyLambda (ExprData closure _ _) ident body) -> do
+  (Eq ann ) =>
+  InterpretFn ann ->
+  ann ->
+  InterpretExpr ann ->
+  InterpretExpr ann ->
+  InterpreterM ann (InterpretExpr ann)
+interpretApp interpretFn ann func arg = do
+  case func of
+    (HOAS.MyLambda _ _ident body) -> do
       -- interpret arg first
-      intValue <- interpretFn value
-      -- add arg to context
-      let newStackFrame = addVarToFrame (varFromIdent ident) intValue closure
-      -- run body with closure + new arg
-      withNewStackFrame newStackFrame (interpretFn body)
-    (MyConstructor ann' modName const') ->
-      MyApp ann (MyConstructor ann' modName const')
-        <$> interpretFn value
+      intArg <- interpretFn arg
+
+      -- run it
+      interpretFn (body intArg)
+    (HOAS.MyRecursiveLambda _ ident (Identifier _ innerRecIdent) body) -> do
+      debugPrettyM "function (recursive)" (fromHOAS func)
+      debugPrettyM "ident" ident
+      debugPrettyM "recIdent" innerRecIdent
+
+      debugPrettyM "the function itself" (fromHOAS arg)
+
+      -- the trick is that we make 'arg' available in context
+      -- so we can reuse it in recursion
+      withVar innerRecIdent arg (interpretFn (body arg) >>=
+          interpretFn)
+    (HOAS.MyConstructor ann' modName const') ->
+      HOAS.MyApp ann (HOAS.MyConstructor ann' modName const')
+        <$> interpretFn arg
     fn -> do
       -- try and resolve it into something we recognise
       intFn <- interpretFn fn
-      if intFn == fn -- if it hasn't changed, we don't want to end up looping so give up and error
+
+      if fromHOAS intFn == fromHOAS fn -- if it hasn't changed, we don't want to end up looping so give up and error
         then do
-          intValue <- interpretFn value
-          -- at least change the value
-          pure (MyApp ann intFn intValue)
-        else interpretFn (MyApp ann intFn value)
+          intArg <- interpretFn arg
+          -- at least change the arg
+          pure (HOAS.MyApp ann intFn intArg)
+        else interpretFn (HOAS.MyApp ann intFn arg)