--- a/compiler/src/Language/Mimsa/Interpreter/App.hs
+++ b/compiler/src/Language/Mimsa/Interpreter/App.hs
@@ -1,6 +1,5 @@
 module Language.Mimsa.Interpreter.App (interpretApp) where
 
-import Language.Mimsa.Logging
 import  Language.Mimsa.Interpreter.Monad
 import qualified Language.Mimsa.Interpreter.HOASExpr as HOAS
 import Language.Mimsa.Interpreter.ToHOAS
@@ -23,12 +22,6 @@ interpretApp interpretFn ann func arg = do
       -- run it
       interpretFn (body intArg)
     (HOAS.MyRecursiveLambda _ ident (Identifier _ innerRecIdent) body) -> do
-      debugPrettyM "function (recursive)" (fromHOAS func)
-      debugPrettyM "ident" ident
-      debugPrettyM "recIdent" innerRecIdent
-
-      debugPrettyM "the function itself" (fromHOAS arg)
-
       -- the trick is that we make 'arg' available in context
       -- so we can reuse it in recursion
       withVar innerRecIdent arg (interpretFn (body arg) >>=