--- a/compiler/src/Language/Mimsa/Interpreter/App.hs
+++ b/compiler/src/Language/Mimsa/Interpreter/App.hs
@@ -1,13 +1,13 @@
 module Language.Mimsa.Interpreter.App (interpretApp) where
 
-import  Language.Mimsa.Interpreter.Monad
+import Language.Mimsa.Core
 import qualified Language.Mimsa.Interpreter.HOASExpr as HOAS
+import Language.Mimsa.Interpreter.Monad
 import Language.Mimsa.Interpreter.ToHOAS
 import Language.Mimsa.Interpreter.Types
-import Language.Mimsa.Core
 
 interpretApp ::
-  (Eq ann ) =>
+  (Eq ann) =>
   InterpretFn ann ->
   ann ->
   InterpretExpr ann ->
@@ -21,11 +21,15 @@ interpretApp interpretFn ann func arg = do
 
       -- run it
       interpretFn (body intArg)
-    (HOAS.MyRecursiveLambda _ ident (Identifier _ innerRecIdent) body) -> do
+    (HOAS.MyRecursiveLambda _ _ident (Identifier _ innerRecIdent) body) -> do
       -- the trick is that we make 'arg' available in context
       -- so we can reuse it in recursion
-      withVar innerRecIdent arg (interpretFn (body arg) >>=
-          interpretFn)
+      withVar
+        innerRecIdent
+        arg
+        ( interpretFn (body arg)
+            >>= interpretFn
+        )
     (HOAS.MyConstructor ann' modName const') ->
       HOAS.MyApp ann (HOAS.MyConstructor ann' modName const')
         <$> interpretFn arg