--- a/compiler/src/Language/Mimsa/Interpreter/If.hs
+++ b/compiler/src/Language/Mimsa/Interpreter/If.hs
@@ -2,27 +2,28 @@ module Language.Mimsa.Interpreter.If (interpretIf) where
 
 import Control.Monad.Except
 import Language.Mimsa.Core
+import qualified Language.Mimsa.Interpreter.HOASExpr as HOAS
+import Language.Mimsa.Interpreter.ToHOAS
 import Language.Mimsa.Interpreter.Types
 import Language.Mimsa.Types.Error.InterpreterError
-import Language.Mimsa.Types.Interpreter.Stack
 
 interpretIf ::
-  InterpretFn var ann ->
-  ExprData var ann ->
-  InterpretExpr var ann ->
-  InterpretExpr var ann ->
-  InterpretExpr var ann ->
-  InterpreterM var ann (InterpretExpr var ann)
+  InterpretFn ann ->
+  ann ->
+  InterpretExpr ann ->
+  InterpretExpr ann ->
+  InterpretExpr ann ->
+  InterpreterM ann (InterpretExpr ann)
 interpretIf interpretFn ann predicate true false =
   case predicate of
-    (MyLiteral _ (MyBool pred')) ->
+    (HOAS.MyLiteral _ (MyBool pred')) ->
       if pred'
         then interpretFn true
         else interpretFn false
-    all'@MyLiteral {} ->
-      throwError $ PredicateForIfMustBeABoolean all'
-    all'@MyLambda {} ->
-      throwError $ PredicateForIfMustBeABoolean all'
+    all'@HOAS.MyLiteral {} ->
+      throwError $ PredicateForIfMustBeABoolean (fromHOAS all')
+    all'@HOAS.MyLambda {} ->
+      throwError $ PredicateForIfMustBeABoolean (fromHOAS all')
     pred' -> do
       predExpr <- interpretFn pred'
-      interpretFn (MyIf ann predExpr true false)
+      interpretFn (HOAS.MyIf ann predExpr true false)