--- a/compiler/src/Language/Mimsa/Interpreter/Infix.hs
+++ b/compiler/src/Language/Mimsa/Interpreter/Infix.hs
@@ -1,34 +1,35 @@
 module Language.Mimsa.Interpreter.Infix (interpretInfix) where
 
 import Control.Monad.Except
-import Language.Mimsa.Core
+import Language.Mimsa.Core hiding (Expr (..))
+import Language.Mimsa.Interpreter.HOASExpr
 import Language.Mimsa.Interpreter.Monad
 import Language.Mimsa.Interpreter.SimpleExpr
+import Language.Mimsa.Interpreter.ToHOAS
 import Language.Mimsa.Interpreter.Types
 import Language.Mimsa.Types.Error.InterpreterError
 
--- | this assumes that
 interpretInfix ::
-  (Ord var, Monoid ann) =>
-  InterpretFn var ann ->
+  (Monoid ann) =>
+  InterpretFn ann ->
   Operator ->
-  InterpretExpr var ann ->
-  InterpretExpr var ann ->
-  InterpreterM var ann (InterpretExpr var ann)
+  InterpretExpr ann ->
+  InterpretExpr ann ->
+  InterpreterM ann (InterpretExpr ann)
 interpretInfix interpretFn operator a b = do
-  plainA <- interpretFn <=< interpretFn $ a
-  plainB <- interpretFn <=< interpretFn $ b
+  plainA <- interpretFn a
+  plainB <- interpretFn b
   case operator of
     Equals -> do
       let withBool = pure . MyLiteral mempty . MyBool
-      if simpleExpr plainA == simpleExpr plainB
+      if simpleExpr (fromHOAS plainA) == simpleExpr (fromHOAS plainB)
         then withBool True
         else withBool False
     Add -> do
       let withInt = pure . MyLiteral mempty . MyInt
       let getNum exp' = case exp' of
             (MyLiteral _ (MyInt i)) -> Right i
-            _ -> Left $ AdditionWithNonNumber a
+            _ -> Left $ AdditionWithNonNumber (fromHOAS exp')
       case (,) <$> getNum plainA <*> getNum plainB of
         Right (a', b') -> withInt (a' + b')
         Left e -> throwError e
@@ -36,32 +37,32 @@ interpretInfix interpretFn operator a b = do
       let withInt = pure . MyLiteral mempty . MyInt
       let getNum exp' = case exp' of
             (MyLiteral _ (MyInt i)) -> Right i
-            _ -> Left $ SubtractionWithNonNumber exp'
+            _ -> Left $ SubtractionWithNonNumber (fromHOAS exp')
       case (,) <$> getNum plainA <*> getNum plainB of
         Right (a', b') -> withInt (a' - b')
         Left e -> throwError e
     GreaterThan ->
       numericComparison
         (>)
-        (ComparisonWithNonNumber GreaterThan)
+        (ComparisonWithNonNumber GreaterThan . fromHOAS)
         plainA
         plainB
     GreaterThanOrEqualTo ->
       numericComparison
         (>=)
-        (ComparisonWithNonNumber GreaterThanOrEqualTo)
+        (ComparisonWithNonNumber GreaterThanOrEqualTo . fromHOAS)
         plainA
         plainB
     LessThan ->
       numericComparison
         (<)
-        (ComparisonWithNonNumber LessThan)
+        (ComparisonWithNonNumber LessThan . fromHOAS)
         plainA
         plainB
     LessThanOrEqualTo ->
       numericComparison
         (<=)
-        (ComparisonWithNonNumber LessThanOrEqualTo)
+        (ComparisonWithNonNumber LessThanOrEqualTo . fromHOAS)
         plainA
         plainB
     StringConcat ->
@@ -80,12 +81,12 @@ interpretInfix interpretFn operator a b = do
 
 -- | lift a numeric comparison into the Expr type
 numericComparison ::
-  (Ord var, Monoid ann) =>
+  (Monoid ann) =>
   (Int -> Int -> Bool) ->
-  (InterpretExpr var ann -> InterpreterError var ann) ->
-  InterpretExpr var ann ->
-  InterpretExpr var ann ->
-  InterpreterM var ann (InterpretExpr var ann)
+  (InterpretExpr ann -> InterpreterError Name ann) ->
+  InterpretExpr ann ->
+  InterpretExpr ann ->
+  InterpreterM ann (InterpretExpr ann)
 numericComparison f withErr plainA plainB = do
   let withBool = pure . MyLiteral mempty . MyBool
   let getNum exp' = case exp' of
@@ -96,29 +97,29 @@ numericComparison f withErr plainA plainB = do
     Left e -> throwError e
 
 interpretStringConcat ::
-  (Ord var, Monoid ann) =>
-  InterpretExpr var ann ->
-  InterpretExpr var ann ->
-  InterpreterM var ann (InterpretExpr var ann)
+  (Monoid ann) =>
+  InterpretExpr ann ->
+  InterpretExpr ann ->
+  InterpreterM ann (InterpretExpr ann)
 interpretStringConcat plainA plainB = do
   let withStr = pure . MyLiteral mempty . MyString . StringType
       getStr exp' = case exp' of
         (MyLiteral _ (MyString (StringType i))) -> Right i
-        _ -> Left $ StringConcatenationFailure plainA plainB
+        _ -> Left $ StringConcatenationFailure (fromHOAS plainA) (fromHOAS plainB)
   case (,) <$> getStr plainA <*> getStr plainB of
     Right (a', b') -> withStr (a' <> b')
     Left e -> throwError e
 
 interpretArrayConcat ::
-  (Ord var, Monoid ann) =>
-  InterpretExpr var ann ->
-  InterpretExpr var ann ->
-  InterpreterM var ann (InterpretExpr var ann)
+  (Monoid ann) =>
+  InterpretExpr ann ->
+  InterpretExpr ann ->
+  InterpreterM ann (InterpretExpr ann)
 interpretArrayConcat plainA plainB = do
   let withArr = pure . MyArray mempty
       getArr exp' = case exp' of
         (MyArray _ i) -> Right i
-        _ -> Left $ ArrayConcatenationFailure plainA plainB
+        _ -> Left $ ArrayConcatenationFailure (fromHOAS plainA) (fromHOAS plainB)
   case (,) <$> getArr plainA <*> getArr plainB of
     Right (a', b') -> withArr (a' <> b')
     Left e -> throwError e