This is mostly great and sometimes broken BUT I think that's because we handle modules in the interpreter stupidly (ie, we should erase them by that point but we don't for some reason).

Old:

```
Running 1 benchmarks...
Benchmark mimsa-benchmark: RUNNING...
benchmarking build stdlib/allFns
time                 148.4 ms   (147.0 ms .. 149.7 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 146.1 ms   (144.9 ms .. 147.2 ms)
std dev              1.692 ms   (1.299 ms .. 2.272 ms)
variance introduced by outliers: 12% (moderately inflated)

benchmarking evaluate/evaluate big looping thing
time                 136.8 ms   (133.4 ms .. 140.8 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 139.6 ms   (137.3 ms .. 143.0 ms)
std dev              4.405 ms   (2.799 ms .. 5.857 ms)
variance introduced by outliers: 11% (moderately inflated)

benchmarking evaluate/evaluate parsing
time                 39.58 ms   (37.79 ms .. 41.94 ms)
                     0.994 R²   (0.988 R² .. 0.999 R²)
mean                 39.06 ms   (38.20 ms .. 40.75 ms)
std dev              2.496 ms   (905.5 μs .. 4.283 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking evaluate/evaluate parsing 2
time                 293.2 ms   (278.1 ms .. 304.6 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 296.4 ms   (293.1 ms .. 300.6 ms)
std dev              4.781 ms   (2.761 ms .. 6.430 ms)
variance introduced by outliers: 16% (moderately inflated)
```


New:
```
Running 1 benchmarks...
Benchmark mimsa-benchmark: RUNNING...
benchmarking build stdlib/allFns
time                 116.7 ms   (115.0 ms .. 121.6 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 114.9 ms   (113.3 ms .. 116.5 ms)
std dev              2.494 ms   (1.563 ms .. 3.486 ms)
variance introduced by outliers: 11% (moderately inflated)

benchmarking evaluate/evaluate big looping thing
time                 83.24 ms   (82.05 ms .. 85.11 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 83.99 ms   (83.47 ms .. 84.44 ms)
std dev              819.3 μs   (565.1 μs .. 1.057 ms)

benchmarking evaluate/evaluate parsing
time                 30.00 ms   (29.67 ms .. 30.54 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 29.97 ms   (29.74 ms .. 30.36 ms)
std dev              646.4 μs   (449.1 μs .. 937.8 μs)

benchmarking evaluate/evaluate parsing 2
time                 40.71 ms   (40.00 ms .. 41.80 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 40.13 ms   (39.85 ms .. 40.78 ms)
std dev              755.1 μs   (368.9 μs .. 1.098 ms)
```