@danieljharvey 2023-03-05
This was so close to good but the resulting code is super hard to understand as soon as recursion gets involved, so abandoning this time sink for now.